pyproject.toml
[tool.poetry]
name = "ai_system_pipeline"
version = "0.1.0"
description = "Monorepo contenente crawler e ingestor"
authors = ["Tuo Nome <tuo@email>"]
packages = [
  { include = "crawler" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = { version = "^0.115" }
crawl4ai = { version = "^0.6.2"}
starlette = { version = "^0.46.2"}
pydantic = { version = "^2.11.4"}

llama-index    = { version = "0.10.3" }
qdrant-client  = { version = "^1.14.2"}
requests       = { version = "^2.32.3" }
openai         = { version = "^1.78.0"}
python-dotenv  = { version = "^1.1.0" }
llama-index-vector-stores-qdrant = { version = ">=0.2.15,<0.3"}
uvicorn = "^0.34.2"
playwright = "^1.52.0"
markdownify = "^1.1.0"
beautifulsoup4 = "^4.13.4"



[tool.poetry.extras]
crawler   = ["fastapi", "uvicorn", "crawl4ai", "playwright", "starlette","pydantic"]
ingestor = [
  "llama-index",
  "llama-index-vector-stores-qdrant",
  "qdrant-client",
  "requests",
  "openai",
  "python-dotenv"
]

[tool.poetry.scripts]

run-ingestor = "ingestor.main:main"
start-crawler = "crawler.start_crawler:main"
retrieve = "retriever.main:main"

[tool.poetry.group.ingestor.dependencies]
llama-index-vector-stores-qdrant = ">=0.2.15,<0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"